1) Added memory profiler   - added memory_profiler to environment.yml   - added profiler functions 2) reorganized the profiler code.
